feat: split big js into small ones
