+++ index.html
@@ -235,6 +235,6 @@
     </footer>
   </main>
 </body>
-<script src="./sctipt.js"></script>
+<script type="module" src="./sctipt.js"></script>
 
 </html>
