+++ pets.html
@@ -117,6 +117,6 @@
     </footer>
   </main>
 </body>
-<script src="./sctipt.js"></script>
+<script type="module" src="./sctipt.js"></script>
 
 </html>
